+++ source/Setup/functions.php
@@ -48,13 +48,11 @@ if (!function_exists('getCountryList')) {
      */
     function getCountryList()
     {
+        $cePath = (new Facts)->getCommunityEditionSourcePath();
         $aCountries = [];
         $relativePath = 'Application/Controller/Admin/ShopCountries.php';
-        if (file_exists(getVendorDirectory() . "oxid-esales/oxideshop-ce/source/$relativePath")) {
-            include getVendorDirectory() . "oxid-esales/oxideshop-ce/source/$relativePath";
-        } else {
-            include __DIR__ . "/../$relativePath";
-        }
+
+        include "$cePath/$relativePath";
 
         return $aCountries;
     }
@@ -68,13 +66,12 @@ if (!function_exists('getLocation')) {
      */
     function getLocation()
     {
+        $cePath = (new Facts)->getCommunityEditionSourcePath();
         $aLocationCountries = [];
         $relativePath = 'Application/Controller/Admin/ShopCountries.php';
-        if (file_exists(getVendorDirectory() . "oxid-esales/oxideshop-ce/source/$relativePath")) {
-            include getVendorDirectory() . "oxid-esales/oxideshop-ce/source/$relativePath";
-        } else {
-            include __DIR__ . "/../$relativePath";
-        }
+
+        include "$cePath/$relativePath";
+
         return $aLocationCountries;
     }
 }
@@ -87,13 +84,11 @@ if (!function_exists('getLanguages')) {
      */
     function getLanguages()
     {
+        $cePath = (new Facts)->getCommunityEditionSourcePath();
         $aLanguages = [];
         $relativePath = 'Application/Controller/Admin/ShopCountries.php';
-        if (file_exists(getVendorDirectory() . "oxid-esales/oxideshop-ce/source/$relativePath")) {
-            include getVendorDirectory() . "oxid-esales/oxideshop-ce/source/$relativePath";
-        } else {
-            include __DIR__ . "/../$relativePath";
-        }
+
+        include "$cePath/$relativePath";
 
         return $aLanguages;
     }